Added source for the eclipse test framework
This source was pulled from a branch of the lombok project. Given the
nature of a eclipse plugin project (which this is) it is hard if
not impossible to build and run this project next to lombok. Pulling
it out and putting it into a separate project is the first step in
making the tests viable again.